Modify `MessageAgent`
